Clearing post-useer relations

+++ controllers/feed.js
@@ -2,7 +2,9 @@ const fs = require('fs');
 const path = require('path');
 
 const { validationResult } = require('express-validator');
+
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -39,6 +41,7 @@ exports.createPost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
     }
+
     if(!req.file) {
         const error = new Error('No image provided');
         error.statusCode = 422;
@@ -48,19 +51,29 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: 'Dan\'SK' },
+        creator: req.userId,
     });
 
     post.save()
         .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'Post created successfully',
-                post: result
+                post: result,
+                creator: { _id: creator._id, name: creator.name }
             })
         })
         .catch(err => {
@@ -110,13 +123,6 @@ exports.updatePost = (req, res, next) => {
     if(req.file) {
         imageUrl = req.file.path;
     }
-    console.log(imageUrl);
-
-    if(!imageUrl) {
-        const error = new Error('No file picked.');
-        error.statusCode = 422;
-        throw error;
-    }
 
     Post.findById(postId)
         .then(post => {
@@ -125,12 +131,19 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
+
             if(imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
             post.title = title;
             post.content = content;
-            post.imageUrl = imageUrl;
+            post.imageUrl = imageUrl ? imageUrl : post.imageUrl;
             return post.save();
         })
         .then(result => {
@@ -156,11 +169,24 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
+
             clearImage(post.imageUrl);
             return Post.deleteOne({ _id: postId });
         })
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
             res.status(200).json({ message: 'Deleted post.' });
         })
         .catch(err => {
